+++ src/kotlin-intellij/test/uk/co/reecedunn/intellij/plugin/core/tests/http/mime/MimeResponseTest.kt
@@ -34,7 +34,7 @@ class MimeResponseTest {
             BasicHeader("Content-Length", "0")
         )
         val body = ""
-        val response = MimeResponse(StringMessage(headers, body), Charsets.ISO_8859_1)
+        val response = MimeResponse(StringMessage(headers, body))
 
         assertThat(response.parts.size, `is`(1))
         assertThat(response.getHeader("Content-Length"), `is`("0"))
@@ -55,7 +55,7 @@ class MimeResponseTest {
             BasicHeader("Content-Type", "text/plain")
         )
         val body = "Hello"
-        val response = MimeResponse(StringMessage(headers, body), Charsets.ISO_8859_1)
+        val response = MimeResponse(StringMessage(headers, body))
 
         assertThat(response.parts.size, `is`(1))
         assertThat(response.getHeader("Content-Length"), `is`("5"))
@@ -85,7 +85,7 @@ class MimeResponseTest {
             "--212ab95a34643c9d--",
             ""
         ).joinToString("\r\n")
-        val response = MimeResponse(StringMessage(headers, body), Charsets.ISO_8859_1)
+        val response = MimeResponse(StringMessage(headers, body))
 
         assertThat(response.parts.size, `is`(1))
         assertThat(response.getHeader("Content-Length"), `is`("98"))
@@ -117,7 +117,7 @@ class MimeResponseTest {
             "--212ab95a34643c9d--",
             ""
         ).joinToString("\r\n")
-        val response = MimeResponse(StringMessage(headers, body), Charsets.ISO_8859_1)
+        val response = MimeResponse(StringMessage(headers, body))
 
         assertThat(response.parts.size, `is`(1))
         assertThat(response.getHeader("Content-Length"), `is`("98"))
@@ -152,7 +152,7 @@ class MimeResponseTest {
             "--47c813e0bbfa09d4--",
             ""
         ).joinToString("\r\n")
-        val response = MimeResponse(StringMessage(headers, body), Charsets.ISO_8859_1)
+        val response = MimeResponse(StringMessage(headers, body))
 
         assertThat(response.parts.size, `is`(2))
         assertThat(response.getHeader("Content-Length"), `is`("205"))
@@ -191,7 +191,7 @@ class MimeResponseTest {
             BasicHeader("Content-Type", "multipart/mixed; boundary=31c406fa29f12029"),
             BasicHeader("X-Content-Type", "text/turtle")
         )
-        val response = MimeResponse(StringMessage(headers, body), Charsets.ISO_8859_1)
+        val response = MimeResponse(StringMessage(headers, body))
 
         assertThat(response.parts.size, `is`(1))
         assertThat(response.getHeader("Content-Length"), `is`("222"))